--- .idea/.name
@@ -1,1 +1,0 @@
-cpa-word-problem
