--- app/src/main/java/com/cpa/cpa_word_problem/data/ProblemContract.kt
@@ -1,17 +1,0 @@
-package com.cpa.cpa_word_problem.data
-
-object ProblemContract {
-    object ProblemEntity {
-        const val TABLE_NAME = "problems"
-        const val PID = "pid"
-        const val YEAR = "year"
-        const val DESCRIPTION = "description"
-        const val P1 = "p1"
-        const val P2 = "p2"
-        const val P3 = "p3"
-        const val P4 = "p4"
-        const val P5 = "p5"
-        const val ANSWER = "answer"
-        const val TYPE = "type"
-    }
-}
