--- app/src/main/java/com/cpa/cpa_word_problem/data/ProblemType.kt
@@ -1,5 +1,0 @@
-package com.cpa.cpa_word_problem.data
-
-enum class ProblemType {
-    Accounting, Business
-}
