--- app/src/main/java/com/cpa/cpa_word_problem/data/QuizOption.kt
@@ -1,4 +1,0 @@
-package com.cpa.cpa_word_problem.data
-
-data class QuizOption(val years: ArrayList<Int>, val type: String) {
-}
