--- app/src/main/java/com/cpa/cpa_word_problem/data/QuizState.kt
@@ -1,5 +1,0 @@
-package com.cpa.cpa_word_problem.data
-
-enum class QuizState {
-    Ready, Start
-}
