+++ app/src/main/java/com/sys0927/cpawordproblem/data/AppDatabase.kt
@@ -1,0 +1,9 @@
+package com.sys0927.cpawordproblem.data
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(entities = [ProblemData::class], version = 2)
+abstract class AppDatabase : RoomDatabase() {
+    abstract fun problemDao(): ProblemDao
+}
