+++ app/src/main/java/com/sys0927/cpawordproblem/data/ProblemType.kt
@@ -1,0 +1,5 @@
+package com.sys0927.cpawordproblem.data
+
+enum class ProblemType {
+    Accounting, Business
+}
