+++ app/src/main/java/com/sys0927/cpawordproblem/data/QuizOption.kt
@@ -1,0 +1,4 @@
+package com.sys0927.cpawordproblem.data
+
+data class QuizOption(val years: ArrayList<Int>, val type: String) {
+}
