+++ app/src/main/java/com/sys0927/cpawordproblem/data/QuizState.kt
@@ -1,0 +1,5 @@
+package com.sys0927.cpawordproblem.data
+
+enum class QuizState {
+    Ready, Start
+}
