--- app/src/test/java/com/cpa/cpa_word_problem/QuizUnitTest.kt
@@ -1,10 +1,0 @@
-package com.cpa.cpa_word_problem
-
-import org.junit.Test
-
-class QuizUnitTest {
-
-    @Test
-    fun testDefaultSetting() {
-    }
-}
