+++ app/src/test/java/com/sys0927/cpawordproblem/QuizUnitTest.kt
@@ -1,0 +1,10 @@
+package com.sys0927.cpawordproblem
+
+import org.junit.Test
+
+class QuizUnitTest {
+
+    @Test
+    fun testDefaultSetting() {
+    }
+}
